+++ advanced/kv/test/registry_test.exs
@@ -2,10 +2,15 @@ defmodule KV.RegistryTest do
   use ExUnit.Case, async: true
 
   setup do
-    # This start_supervised method runs the KV.Registry start_link/1 method
-    # This also guarantee that the registry process will be shutdown before the next test starts
     registry = start_supervised!(KV.Registry)
     %{registry: registry}
+  end
+
+  test "Testing the monitored feature of the RegistryMonitored", %{registry: registry} do
+    KV.Registry.create(registry, "shopping")
+    {:ok, bucket} = KV.Registry.lookup(registry, "shopping")
+    Agent.stop(bucket)
+    assert KV.Registry.lookup(registry, "shopping") == :error
   end
 
   test "Testing buckets with GenServer", %{registry: registry} do
@@ -19,5 +24,14 @@ defmodule KV.RegistryTest do
     # After below, it uses KV.Bucket to put a value and get it
     KV.Bucket.put(bucket, "milk", 1)
     assert KV.Bucket.get_by_key(bucket, "milk") == 1
+  end
+
+  test "removes bucket on crash", %{registry: registry} do
+    KV.Registry.create(registry, "shopping")
+    {:ok, bucket} = KV.Registry.lookup(registry, "shopping")
+
+    # Stop the bucket with non-normal reason
+    Agent.stop(bucket, :shutdown)
+    assert KV.Registry.lookup(registry, "shopping") == :error
   end
 end
